requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
streamlit==1.40.0
sqlite-utils==3.35.1
python-dotenv
langchain==0.3.27
langchain-community==0.3.29
langchain-text-splitters==0.3.11
langchain-huggingface==0.3.1
huggingface-hub>=0.33.4
sentence-transformers==3.0.0
langchain-chroma==0.2.6
chromadb==1.5.1
langchain-google-genai==2.0.10
google-generativeai==0.8.3
jinja2
aiofiles
python-multipart